+++ src/cpu.h
@@ -19,7 +19,7 @@ class Access {
 
 
 
-int ModuloWidth(uint32_t addr, int bit_width, int pos);
+unsigned int ModuloWidth(uint64_t addr, unsigned int bit_width, unsigned int pos);
 Address AddressMapping(uint64_t hex_addr, const Config& config);
